start to remove core from daemon
