+++ consensus/transactions_test.go
@@ -7,7 +7,7 @@ import (
 // TestApplyTransaction provides testing coverage for State.applyTransaction()
 func TestApplyTransaction(t *testing.T) {
 	// Create a state to which transactions can be applied.
-	s, _ := CreateGenesisState()
+	s := CreateGenesisState()
 
 	// Create a transaction with one input and one output.
 	transaction := Transaction{
