+++ siad/apihost.go
@@ -1,52 +1,58 @@
 package main
 
 import (
-	"fmt"
+	// "fmt"
 	"net/http"
 
-	"github.com/NebulousLabs/Sia/sia/components"
+	// "github.com/NebulousLabs/Sia/sia/components"
 )
 
 // Provides the configuration settings for the host.
 func (d *daemon) hostConfigHandler(w http.ResponseWriter, req *http.Request) {
-	hinfo, err := d.core.HostInfo()
-	if err != nil {
-		http.Error(w, "Failed to retreive HostInfo: "+err.Error(), 405)
-	}
+	http.Error(w, "Unimplemented", 500)
+	/*
+		hinfo, err := d.core.HostInfo()
+		if err != nil {
+			http.Error(w, "Failed to retreive HostInfo: "+err.Error(), 405)
+		}
 
-	writeJSON(w, hinfo)
+		writeJSON(w, hinfo)
+	*/
 }
 
 func (d *daemon) hostSetConfigHandler(w http.ResponseWriter, req *http.Request) {
-	hAnnouncement := components.HostAnnouncement{}
+	http.Error(w, "Unimplemented", 500)
+	/*
+		hAnnouncement := components.HostAnnouncement{}
 
-	qsVars := map[string]interface{}{
-		"totalstorage": &hAnnouncement.TotalStorage,
-		// "minfile":      &hAnnouncement.MinFilesize,
-		"maxfilesize": &hAnnouncement.MaxFilesize,
-		// "minduration":  &hAnnouncement.MinDuration,
-		"maxduration": &hAnnouncement.MaxDuration,
-		"price":       &hAnnouncement.Price,
-		"burn":        &hAnnouncement.Burn,
-	}
-
-	for qs := range qsVars {
-		_, err := fmt.Sscan(req.FormValue(qs), qsVars[qs])
-		if err != nil {
-			http.Error(w, "Malformed "+qs+" "+err.Error(), 400)
-			return
+		qsVars := map[string]interface{}{
+			"totalstorage": &hAnnouncement.TotalStorage,
+			// "minfile":      &hAnnouncement.MinFilesize,
+			"maxfilesize": &hAnnouncement.MaxFilesize,
+			// "minduration":  &hAnnouncement.MinDuration,
+			"maxduration": &hAnnouncement.MaxDuration,
+			"price":       &hAnnouncement.Price,
+			"burn":        &hAnnouncement.Burn,
 		}
-	}
 
-	err := d.core.UpdateHost(hAnnouncement)
-	if err != nil {
-		http.Error(w, "Could not update host:"+err.Error(), 400)
-	}
+		for qs := range qsVars {
+			_, err := fmt.Sscan(req.FormValue(qs), qsVars[qs])
+			if err != nil {
+				http.Error(w, "Malformed "+qs+" "+err.Error(), 400)
+				return
+			}
+		}
 
-	err = d.core.AnnounceHost(1, d.state.Height()+20) // A freeze volume and unlock height.
-	if err != nil {
-		http.Error(w, "Could not update host:"+err.Error(), 400)
-	}
+		err := d.core.UpdateHost(hAnnouncement)
+		if err != nil {
+			http.Error(w, "Could not update host:"+err.Error(), 400)
+		}
 
-	writeSuccess(w)
+		err = d.core.AnnounceHost(1, d.state.Height()+20) // A freeze volume and unlock height.
+		if err != nil {
+			http.Error(w, "Could not update host:"+err.Error(), 400)
+		}
+
+		writeSuccess(w)
+	*/
 }
