+++ siad/daemon_test.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestStartDaemon(t *testing.T) {
+	// Do nothing
+}
